requirements.txt
pyTelegramBotAPI
requests
pycryptodome
python-dotenv
